requirements.txt
streamlit
pandas
joblib
faiss-cpu
